Moved dataloader out of torch
