--- benchmarks/rnnt/ootb/train/common/data/torchaudio/data_loader.py
@@ -1,497 +1,0 @@
-import os
-import subprocess
-from tempfile import NamedTemporaryFile
-
-from torch.distributed import get_rank
-from torch.distributed import get_world_size
-# from torch.distributed.deprecated import get_rank
-# from torch.distributed.deprecated import get_world_size
-from torch.utils.data.sampler import Sampler
-
-import librosa
-import random
-import numpy as np
-import scipy.signal
-import torch
-import torch.functional as F
-import torchaudio
-import math
-from torch.utils.data import DataLoader
-from torch.utils.data import Dataset
-from common.data.dataset import AudioDataset
-
-#TODO This load fails, but we don't need it anyways, only a small part of this file is used anymore(AudioDataLoader).  We should move it.
-#from data.SpecAugment import sparse_image_warp_zcaceres
-
-windows = {'hamming': scipy.signal.hamming, 'hann': scipy.signal.hann, 'blackman': scipy.signal.blackman,
-           'bartlett': scipy.signal.bartlett}
-
-def mel_filter_bank(inpus):
-    low_freq_mel = 0
-    high_freq_mel = (2595 * numpy.log10(1 + (sample_rate / 2) / 700))  # Convert Hz to Mel
-    mel_points = numpy.linspace(low_freq_mel, high_freq_mel, nfilt + 2)  # Equally spaced in Mel scale
-    hz_points = (700 * (10 ** (mel_points / 2595) - 1))  # Convert Mel to Hz
-    bin = numpy.floor((NFFT + 1) * hz_points / sample_rate)
-
-    fbank = numpy.zeros((nfilt, int(numpy.floor(NFFT / 2 + 1))))
-    for m in range(1, nfilt + 1):
-        f_m_minus = int(bin[m - 1])  # left
-        f_m = int(bin[m])  # center
-        f_m_plus = int(bin[m + 1])  # right
-
-        for k in range(f_m_minus, f_m):
-            fbank[m - 1, k] = (k - bin[m - 1]) / (bin[m] - bin[m - 1])
-        for k in range(f_m, f_m_plus):
-            fbank[m - 1, k] = (bin[m + 1] - k) / (bin[m + 1] - bin[m])
-    filter_banks = numpy.dot(pow_frames, fbank.T)
-    filter_banks = numpy.where(filter_banks == 0, numpy.finfo(float).eps, filter_banks)  # Numerical Stability
-    filter_banks = 20 * numpy.log10(filter_banks)  # dB
-
-    return filter_banks
-
-
-def zero_pad_concat(inputs):
-    max_t = max(len(inp) for inp in inputs)
-    shape = (len(inputs), max_t) + inputs[0].shape[1:]
-    input_mat = np.zeros(shape, dtype=np.float32)
-    for e, inp in enumerate(inputs):
-        input_mat[e, :inp.shape[0]] = inp
-    return input_mat
-
-
-def end_pad_label(inputs):
-    max_t = max(len(i) for i in inputs)
-    # max_t = 50
-    shape = (len(inputs), max_t)
-    labels = np.full(shape, fill_value=26, dtype='i')
-    for e, l in enumerate(inputs):
-        labels[e, :len(l)] = l
-
-    return labels
-
-
-def time_warp(spec, W=3):
-    spec = spec.view(1, spec.shape[0], spec.shape[1])
-    num_rows = spec.shape[1]
-    spec_len = spec.shape[2]
-
-    y = num_rows // 2
-    horizontal_line_at_ctr = spec[0][y]
-    assert len(horizontal_line_at_ctr) == spec_len
-
-    point_to_warp = horizontal_line_at_ctr[random.randrange(W, spec_len - W)]
-    assert isinstance(point_to_warp, torch.Tensor)
-
-    # Uniform distribution from (0,W) with chance to be up to W negative
-    dist_to_warp = random.randrange(-W, W)
-    src_pts, dest_pts = torch.tensor([[[y, point_to_warp]]]), torch.tensor([[[y, point_to_warp + dist_to_warp]]])
-    warped_spectro, dense_flows = sparse_image_warp_zcaceres.sparse_image_warp(spec, src_pts, dest_pts)
-    return warped_spectro.squeeze(3)
-
-
-def freq_mask(spec, F=15, num_masks=1, replace_with_zero=True):
-    cloned = spec.clone()
-    num_mel_channels = cloned.shape[1]
-
-    for i in range(0, num_masks):
-        f = random.randrange(0, F)
-        f_zero = random.randrange(0, num_mel_channels - f)
-
-        # avoids randrange error if values are equal and range is empty
-        if (f_zero == f_zero + f): return cloned
-
-        mask_end = random.randrange(f_zero, f_zero + f)
-        if (replace_with_zero):
-            cloned[0][f_zero:mask_end] = 0
-        else:
-            cloned[0][f_zero:mask_end] = cloned.mean()
-
-    return cloned
-
-
-def time_mask(spec, T=15, num_masks=1, replace_with_zero=True):
-    cloned = spec.clone()
-    len_spectro = cloned.shape[2]
-
-    for i in range(0, num_masks):
-        t = random.randrange(0, T)
-        t_zero = random.randrange(0, len_spectro - t)
-
-        # avoids randrange error if values are equal and range is empty
-        if (t_zero == t_zero + t): return cloned
-
-        mask_end = random.randrange(t_zero, t_zero + t)
-        if (replace_with_zero):
-            cloned[0][:, t_zero:mask_end] = 0
-        else:
-            cloned[0][:, t_zero:mask_end] = cloned.mean()
-    return cloned
-
-
-def load_audio(path):
-    sound, _ = torchaudio.load(path)
-    sound = sound.numpy().T
-    if len(sound.shape) > 1:
-        if sound.shape[1] == 1:
-            sound = sound.squeeze()
-        else:
-            sound = sound.mean(axis=1)  # multiple channels, average
-    return sound
-
-
-class AudioParser(object):
-    def parse_transcript(self, transcript_path):
-        """
-        :param transcript_path: Path where transcript is stored from the manifest file
-        :return: Transcript in training/testing format
-        """
-        raise NotImplementedError
-
-    def parse_audio(self, audio_path):
-        """
-        :param audio_path: Path where audio is stored from the manifest file
-        :return: Audio in training/testing format
-        """
-        raise NotImplementedError
-
-
-class NoiseInjection(object):
-    def __init__(self,
-                 path=None,
-                 sample_rate=16000,
-                 noise_levels=(0, 0.5)):
-        """
-        Adds noise to an input signal with specific SNR. Higher the noise level, the more noise added.
-        Modified code from https://github.com/willfrey/audio/blob/master/torchaudio/transforms.py
-        """
-        if not os.path.exists(path):
-            print("Directory doesn't exist: {}".format(path))
-            raise IOError
-        self.paths = path is not None and librosa.util.find_files(path)
-        self.sample_rate = sample_rate
-        self.noise_levels = noise_levels
-
-    def inject_noise(self, data):
-        noise_path = np.random.choice(self.paths)
-        noise_level = np.random.uniform(*self.noise_levels)
-        return self.inject_noise_sample(data, noise_path, noise_level)
-
-    def inject_noise_sample(self, data, noise_path, noise_level):
-        noise_len = get_audio_length(noise_path)
-        data_len = len(data) / self.sample_rate
-        noise_start = np.random.rand() * (noise_len - data_len)
-        noise_end = noise_start + data_len
-        noise_dst = audio_with_sox(noise_path, self.sample_rate, noise_start, noise_end)
-        assert len(data) == len(noise_dst)
-        noise_energy = np.sqrt(noise_dst.dot(noise_dst) / noise_dst.size)
-        data_energy = np.sqrt(data.dot(data) / data.size)
-        data += noise_level * noise_dst * data_energy / noise_energy
-        return data
-
-
-class SpectrogramParser(AudioParser):
-    def __init__(self, audio_conf, normalize=False, augment=False, specaugment=False, mel_filterbank=False):
-        """
-        Parses audio file into spectrogram with optional normalization and various augmentations
-        :param audio_conf: Dictionary containing the sample rate, window and the window length/stride in seconds
-        :param normalize(default False):  Apply standard mean and deviation normalization to audio tensor
-        :param augment(default False):  Apply random tempo and gain perturbations
-        """
-        super(SpectrogramParser, self).__init__()
-        self.window_stride = audio_conf['window_stride']
-        self.window_size = audio_conf['window_size']
-        self.sample_rate = audio_conf['sample_rate']
-        self.window = windows.get(audio_conf['window'], windows['hamming'])
-        self.normalize = normalize
-        self.augment = augment
-        self.noiseInjector = NoiseInjection(audio_conf['noise_dir'], self.sample_rate,
-                                            audio_conf['noise_levels']) if audio_conf.get(
-            'noise_dir') is not None else None
-        self.noise_prob = audio_conf.get('noise_prob')
-        self.specaugment = specaugment
-        self.mel_filterbank = mel_filterbank
-
-    def parse_audio(self, audio_path):
-        if self.augment:
-            y = load_randomly_augmented_audio(audio_path, self.sample_rate)
-        else:
-            y = load_audio(audio_path)
-        if self.noiseInjector:
-            add_noise = np.random.binomial(1, self.noise_prob)
-            if add_noise:
-                y = self.noiseInjector.inject_noise(y)
-        n_fft = int(self.sample_rate * self.window_size)
-        win_length = n_fft
-        hop_length = int(self.sample_rate * self.window_stride)
-        if self.mel_filterbank is True :
-            librosa.filters.mel(sr=self.sample_rate, n_fft=n_fft,
-                                nmels=128)
-        else:
-            # STFT
-            D = librosa.stft(y, n_fft=n_fft, hop_length=hop_length,
-                            win_length=win_length, window=self.window)
-            spect, phase = librosa.magphase(D)
-            # S = log(S+1)
-            spect = np.log1p(spect)
-            spect = torch.FloatTensor(spect)
-
-        if self.specaugment:
-            spect = time_mask(freq_mask(time_warp(spect), num_masks=2), num_masks=2)
-            spect = spect.view(spect.shape[1],spect.shape[2])
-        if self.normalize:
-            mean = spect.mean()
-            std = spect.std()
-            spect.add_(-mean)
-            spect.div_(std)
-
-        # print(spect.size())
-
-        return spect
-
-    def parse_transcript(self, transcript_path):
-        raise NotImplementedError
-
-
-class SpectrogramDataset(Dataset, SpectrogramParser):
-    def __init__(self, audio_conf, manifest_filepath, labels, normalize=False, augment=False, specaugment=False):
-        """
-        Dataset that loads tensors via a csv containing file paths to audio files and transcripts separated by
-        a comma. Each new line is a different sample. Example below:
-
-        /path/to/audio.wav,/path/to/audio.txt
-        ...
-
-        :param audio_conf: Dictionary containing the sample rate, window and the window length/stride in seconds
-        :param manifest_filepath: Path to manifest csv as describe above
-        :param labels: String containing all the possible characters to map to
-        :param normalize: Apply standard mean and deviation normalization to audio tensor
-        :param augment(default False):  Apply random tempo and gain perturbations
-        """
-
-        with open(manifest_filepath) as f:
-            ids = f.readlines()
-        ids = [x.strip().split(',') for x in ids]
-        self.ids = ids
-        self.size = len(ids)
-        self.labels_map = dict([(labels[i], i) for i in range(len(labels))])
-        super(SpectrogramDataset, self).__init__(audio_conf, normalize, augment, specaugment)
-
-    def __getitem__(self, index):
-        sample = self.ids[index]
-        audio_path, transcript_path = sample[0], sample[1]
-        spect = self.parse_audio(audio_path)
-        transcript = self.parse_transcript(transcript_path)
-        # print(spect.size(), len(transcript))
-        transcript_one_hot = torch.nn.functional.one_hot(torch.LongTensor(transcript), num_classes=len(self.labels_map))
-        return spect, transcript, transcript_one_hot, self.labels_map
-
-    def parse_transcript(self, transcript_path):
-        with open(transcript_path, 'r', encoding='utf8') as transcript_file:
-            transcript = transcript_file.read().replace('\n', '')
-        transcript = list(filter(None, [self.labels_map.get(x) for x in list(transcript)]))
-        return transcript
-
-    def __len__(self):
-        return self.size
-
-
-def _collate_fn(batch):
-    bs = len(batch)
-    max_len = lambda l, idx: max(el[idx].size(0) for el in l)
-    max_len2 = lambda l, idx: max(el[0][idx].size(0) for el in l)
-    #TODO, don't use 80...maybe get shape of 2nd dimension of batch
-    audio = torch.zeros(bs, 80, max_len2(batch, 0))
-
-    audio_lens = torch.zeros(bs, dtype=torch.int32)
-    transcript = torch.zeros(bs, max_len(batch, 2))
-    transcript_lens = torch.zeros(bs, dtype=torch.int32)
-
-    for i, sample in enumerate(batch):
-        audio[i].narrow(1, 0, sample[0].size(1)).copy_(sample[0])
-        audio_lens[i] = sample[1]
-        transcript[i].narrow(0, 0, sample[2].size(0)).copy_(sample[2])
-        transcript_lens[i] = sample[3]
-
-    return audio, audio_lens, transcript, transcript_lens
-
-
-class AudioDataLoader(DataLoader):
-    def __init__(self,
-                 #config
-                 config_features,
-                 # DataSet
-                 data_dir, manifest_fpaths,
-                 tokenizer,
-                 sample_rate=16000, min_duration=0.1, max_duration=float("inf"),
-                 max_utts=0, normalize_transcripts=True,
-                 trim_silence=False,
-                 speed_perturbation=None,
-                 ignore_offline_speed_perturbation=False,
-                 # Sampler
-                 batch_size=1, num_replicas=None, rank=None,
-                 # Loader
-                 shuffle=False, sampler=None,
-                 batch_sampler=None, num_workers=0, collate_fn=None,
-                 pin_memory=False, drop_last=False, timeout=0,
-                 worker_init_fn=None, *, prefetch_factor=2,
-                 persistent_workers=False):
-        """
-        Creates a data loader for AudioDatasets.
-        """
-        self.audio_dataset = AudioDataset(
-                 data_dir=data_dir,
-                 tokenizer=tokenizer,
-                 manifest_fpaths=manifest_fpaths,
-                 n_filt=config_features["n_filt"],
-                 n_fft=config_features["n_fft"]
-                 )
-
-        if batch_sampler is None:
-            if num_replicas <= 1:
-                self.batch_sampler = BucketingSampler(self.audio_dataset,
-                                           batch_size=batch_size)
-            else:
-                self.batch_sampler = DistributedBucketingSampler(self.audio_dataset,
-                                            batch_size=batch_size,
-                                            num_replicas=num_replicas,
-                                            rank=rank)
-        else:
-            self.batch_sampler = batch_sampler
-
-        super(AudioDataLoader, self).__init__(self.audio_dataset,
-                                            batch_sampler=self.batch_sampler,
-                                            num_workers=num_workers, collate_fn=None,
-                                            pin_memory=pin_memory,
-                                            timeout=timeout, worker_init_fn=worker_init_fn,
-                                            prefetch_factor=prefetch_factor,
-                                            persistent_workers=persistent_workers)
-        self.collate_fn = _collate_fn
-
-
-class BucketingSampler(Sampler):
-    def __init__(self, data_source, batch_size=1):
-        """
-        Samples batches assuming they are in order of size to batch similarly sized samples together.
-        """
-        super(BucketingSampler, self).__init__(data_source)
-        self.data_source = data_source
-        ids = list(range(0, len(data_source)))
-        self.bins = [ids[i:i + batch_size] for i in range(0, len(ids), batch_size)]
-
-    def __iter__(self):
-        for ids in self.bins:
-            np.random.shuffle(ids)
-            yield ids
-
-    def __len__(self):
-        return len(self.bins)
-
-    def shuffle(self, epoch):
-        np.random.shuffle(self.bins)
-
-
-class BatchRandomSampler(Sampler):
-    """
-    Batches the data consecutively and randomly samples
-    by batch without replacement.
-    """
-
-    def __init__(self, data_source, batch_size):
-        it_end = len(data_source) - batch_size + 1
-        self.batches = [range(i, i + batch_size)
-                for i in range(0, it_end, batch_size)]
-        self.data_source = data_source
-
-    def __iter__(self):
-        random.shuffle(self.batches)
-        return (i for b in self.batches for i in b)
-
-    def __len__(self):
-        return len(self.data_source)
-
-
-class DistributedBucketingSampler(Sampler):
-    def __init__(self, data_source, batch_size=1, num_replicas=None, rank=None):
-        """
-        Samples batches assuming they are in order of size to batch similarly sized samples together.
-        """
-        super(DistributedBucketingSampler, self).__init__(data_source)
-        if num_replicas is None:
-            num_replicas = get_world_size()
-        if rank is None:
-            rank = get_rank()
-        self.data_source = data_source
-        self.ids = list(range(0, len(data_source)))
-        self.batch_size = batch_size
-        self.bins = [self.ids[i:i + batch_size] for i in range(0, len(self.ids), batch_size)]
-        self.num_replicas = num_replicas
-        self.rank = rank
-        self.num_samples = int(math.ceil(len(self.bins) * 1.0 / self.num_replicas))
-        self.total_size = self.num_samples * self.num_replicas
-
-    def __iter__(self):
-        offset = self.rank
-        # add extra samples to make it evenly divisible
-        bins = self.bins + self.bins[:(self.total_size - len(self.bins))]
-        assert len(bins) == self.total_size
-        samples = bins[offset::self.num_replicas]  # Get every Nth bin, starting from rank
-        return iter(samples)
-
-    def __len__(self):
-        return self.num_samples
-
-    def shuffle(self, epoch):
-        # deterministically shuffle based on epoch
-        g = torch.Generator()
-        g.manual_seed(epoch)
-        bin_ids = list(torch.randperm(len(self.bins), generator=g))
-        self.bins = [self.bins[i] for i in bin_ids]
-
-
-def get_audio_length(path):
-    output = subprocess.check_output(['soxi -D \"%s\"' % path.strip()], shell=True)
-    return float(output)
-
-
-def audio_with_sox(path, sample_rate, start_time, end_time):
-    """
-    crop and resample the recording with sox and loads it.
-    """
-    with NamedTemporaryFile(suffix=".wav") as tar_file:
-        tar_filename = tar_file.name
-        sox_params = "sox \"{}\" -r {} -c 1 -b 16 -e si {} trim {} ={} >/dev/null 2>&1".format(path, sample_rate,
-                                                                                               tar_filename, start_time,
-                                                                                               end_time)
-        os.system(sox_params)
-        y = load_audio(tar_filename)
-        return y
-
-
-def augment_audio_with_sox(path, sample_rate, tempo, gain):
-    """
-    Changes tempo and gain of the recording with sox and loads it.
-    """
-    with NamedTemporaryFile(suffix=".wav") as augmented_file:
-        augmented_filename = augmented_file.name
-        sox_augment_params = ["tempo", "{:.3f}".format(tempo), "gain", "{:.3f}".format(gain)]
-        sox_params = "sox \"{}\" -r {} -c 1 -b 16 -e si {} {} >/dev/null 2>&1".format(path, sample_rate,
-                                                                                      augmented_filename,
-                                                                                      " ".join(sox_augment_params))
-        os.system(sox_params)
-        y = load_audio(augmented_filename)
-        return y
-
-
-def load_randomly_augmented_audio(path, sample_rate=16000, tempo_range=(0.85, 1.15),
-                                  gain_range=(-6, 8)):
-    """
-    Picks tempo and gain uniformly, applies it to the utterance by using sox utility.
-    Returns the augmented utterance.
-    """
-    low_tempo, high_tempo = tempo_range
-    tempo_value = np.random.uniform(low=low_tempo, high=high_tempo)
-    low_gain, high_gain = gain_range
-    gain_value = np.random.uniform(low=low_gain, high=high_gain)
-    audio = augment_audio_with_sox(path=path, sample_rate=sample_rate,
-                                   tempo=tempo_value, gain=gain_value)
-    return audio
