+++ benchmarks/rnnt/ootb/train/train.py
@@ -357,7 +357,7 @@ def main():
                                     device_type=args.dali_device,
                                     tokenizer=tokenizer)
     else:
-        from common.data.torchaudio.data_loader import AudioDataLoader
+        from common.data.data_loader import AudioDataLoader
 
         #OUTPUT: tokenizer: {'labels': [' ', 'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z', "'"], 'sentpiece_model': '/data/rnnt/datasets/sentencepieces/librispeech1023.model'}
         print("tokenizer: {}".format(tokenizer_kw))
